+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Addition
+* Added tooltips to arrows
+
 ## 1.18.1
 ### Changes
 * Shields+ is no longer a required dependency!